backend/requirements.txt
openai==0.28
python-dotenv
fastapi
uvicorn[standard]
duckduckgo_search
transformers
torch